requirements.txt
Flask==3.1.3
pandas==3.0.2
scikit-learn==1.8.0
joblib==1.5.3
numpy==2.4.4
python-dotenv==1.0.0
